% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate
This function implements a calibration algorithm based on the Direct Search method (DDS).}
\arguments{
\item{fitness}{An Rcpp-exported C++ function (callable from R and C++) that accepts (x, other_data) and returns arma::vec (first element is objective).}

\item{lst_OtherData}{A field of additional data passed to `fitness`.}

\item{x_Min}{A numeric vector of lower bounds for each parameter.}

\item{x_Max}{A numeric vector of upper bounds for each parameter.}

\item{x_Init}{An optional initial solution (numeric vector). If empty, midpoint of x_Min and x_Max is used.}

\item{max_iter}{Maximum iterations (default 100).}

\item{r}{Perturbation factor (default 0.2).}
}
\value{
arma::vec: best parameter set found.
}
\description{
It attempts to find the optimal parameter values that minimize the given objective (fitness) function.
}
