% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroGalleryCpp-package.R
\docType{_PACKAGE}
\name{HydroGalleryCpp-package}
\alias{HydroGalleryCpp-package}
\title{A Brief Introduction to HydroGalleryCpp}
\description{
An pure C++ core package designed to provide a versatile and efficient suite of functions for hydrological modeling and analysis.
}
