% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meteo}
\alias{meteo}
\alias{meteo_solarRadiatClearSky_FAO56}
\alias{meteo_solarRadiatClearSky_WaterGAP3}
\alias{meteo_saturatVaporPress}
\alias{meteo_saturatVaporPress_kPa}
\alias{meteo_vaporPress}
\alias{meteo_nettoRadiat_FAO56}
\alias{meteo_atmosEmissivity_FAO56}
\alias{meteo_cloudFactor_UNKNOW}
\alias{meteo_atmosEmissivity_UNKNOW}
\alias{meteo_atmosEmissivity_Idso}
\alias{meteo_nettoLongRadiat_WaterGAP3}
\alias{meteo_nettoRadiat_WaterGAP3}
\alias{meteo_nettoRadiat_FAO56Simplify}
\alias{meteo_windSpeed2m}
\alias{meteo_airDensity}
\alias{meteo_saturatDelta}
\alias{meteo_wetBulbTemperature}
\title{**meteological variables**
some functions to calculate the meteological variables}
\usage{
meteo_solarRadiatClearSky_FAO56(
  Time_dayOfYear_,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_solarRadiatClearSky_WaterGAP3(Time_dayOfYear_, LAND_latitude_Degree)

meteo_saturatVaporPress(ATMOS_temperature_Cel)

meteo_saturatVaporPress_kPa(ATMOS_temperature_Cel)

meteo_vaporPress(ATMOS_temperature_Cel, ATMOS_relativeHumidity_1)

meteo_nettoRadiat_FAO56(
  Time_dayOfYear_,
  ATMOS_temperature_Cel,
  ATMOS_temperatureMax_Cel,
  ATMOS_temperatureMin_Cel,
  ATMOS_relativeHumidity_1,
  ATMOS_solarRadiat_MJ,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_atmosEmissivity_FAO56(
  Time_dayOfYear_,
  ATMOS_temperature_Cel,
  ATMOS_relativeHumidity_1,
  ATMOS_solarRadiat_MJ,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_cloudFactor_UNKNOW(
  ATMOS_solarRadiat_MJ,
  Time_dayOfYear_,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_atmosEmissivity_UNKNOW(
  Time_dayOfYear_,
  ATMOS_temperature_Cel,
  ATMOS_vaporPress_kPa,
  ATMOS_solarRadiat_MJ,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_atmosEmissivity_Idso(ATMOS_temperature_Cel)

meteo_nettoLongRadiat_WaterGAP3(
  ATMOS_temperature_Cel,
  ATMOS_solarRadiat_MJ,
  ATMOS_solarRadiatClearSky_MJ,
  param_ATMOS_nettoRadiat_wg3a,
  param_ATMOS_nettoRadiat_wg3b
)

meteo_nettoRadiat_WaterGAP3(
  ATMOS_solarRadiat_MJ,
  ATMOS_nettoLongRadiat_MJ,
  LAND_albedo_1
)

meteo_nettoRadiat_FAO56Simplify(
  Time_dayOfYear_,
  ATMOS_temperature_Cel,
  ATMOS_relativeHumidity_1,
  ATMOS_solarRadiat_MJ,
  LAND_latitude_Degree,
  LAND_elevation_m
)

meteo_windSpeed2m(ATMOS_windSpeed_m_s, ATMOS_windMeasureHeight_m)

meteo_airDensity(ATMOS_temperature_Cel, LAND_elevation_m)

meteo_saturatDelta(ATMOS_temperature_Cel)

meteo_wetBulbTemperature(ATMOS_vaporPress_kPa, ATMOS_temperature_Cel)
}
\arguments{
\item{LAND_elevation_m}{(m) average elevation}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_relativeHumidity_1}{(0, 1) relative humidity}

\item{ATMOS_temperatureMax_Cel}{(Cel) the maximal air temperature in the time phase}

\item{ATMOS_temperatureMin_Cel}{(Cel) the minimal air temperature in the time phase}

\item{ATMOS_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{LAND_albedo_1}{<0, 1> albedo of the region}

\item{ATMOS_windSpeed_m_s}{(m/s) measured wind speed at z m above ground surface}

\item{ATMOS_windMeasureHeight_m}{(m) height of measurement above ground surface}
}
\value{
meteological variables
}
\description{
**meteological variables**
some functions to calculate the meteological variables
}
