% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{routingtopology}
\alias{routingtopology}
\alias{get_inflow_cells}
\alias{get_inflow_lastcell}
\alias{get_step_cells}
\alias{get_step_lastcell}
\alias{generate_step_cell}
\alias{get_step_extra_cell}
\alias{generate_step_extra_cell}
\alias{get_cell_in_basin}
\alias{get_inter_basin}
\alias{get_new_outflow}
\alias{get_cali_step}
\alias{get_upstream_cali_cell}
\title{Get Inflow Cells}
\usage{
get_inflow_cells(int_Outflow)

get_inflow_lastcell(int_Outflow)

get_step_cells(inflow_cells)

get_step_lastcell(step_cells, inflow_lastcell)

generate_step_cell(int_Outflow, fn_Step_Cell, fn_Step_LastCell)

get_step_extra_cell(Step_cellNumber_int, Extra_cellNumber_int)

generate_step_extra_cell(fn_Step_Cell, fn_Extra_Cell, fn_Step_Extra_Cell)

get_cell_in_basin(lst_Inflow_Cell, int_OutLet, int_TestCell)

get_inter_basin(int_Cell, int_Outflow)

get_new_outflow(int_Cell, int_Outflow)

get_cali_step(step_cells, int_Cali)

get_upstream_cali_cell(lst_Inflow_Cell, int_CaliCell)
}
\arguments{
\item{int_Outflow}{A vector of integers representing the cell number of the next cell. 1-based indexing. When the next cell is see or none, should be marked as 0.}

\item{fn_Step_Cell}{Path to read the step_cells field.}

\item{fn_Step_LastCell}{Path to save the step_lastcell field.}

\item{Step_cellNumber_int}{Cell number of the step.}

\item{Extra_cellNumber_int}{Cell number of the extra cells.}

\item{fn_Extra_Cell}{Path to read the extra cells field.}

\item{fn_Step_Extra_Cell}{path to write the step extra cell number.}

\item{lst_Inflow_Cell}{A list of integer vectors, where each vector contains the cells that flow into the respective cell.}

\item{int_OutLet}{An integer representing the outlet cell (1-based index).}

\item{int_TestCell}{An integer vector, cells to test.}

\item{int_CaliCell}{An integer vector of calibration cells.}

\item{int_UpstreamCell}{An integer vector containing the upstream cells to find the upstream basin.}

\item{int_Outflow_Ori}{An integer vector representing the original outflow indices (1-based).}

\item{int_CellNew}{An integer vector representing the cells within the new basin.}
}
\value{
A field of uvecs containing the inflow cells for each cell.

NULL (invisible)

NULL (invisible)

An integer vector of cells in the intersection of the station cells and the basin.

An integer vector representing the new upstream basin, which includes the upstream cells and the set difference of the basin cells.
This function identifies the upstream basin of a given outlet cell by first finding the intersection of the upstream cells
with the cells that flow into the outlet. It then computes the set difference between the upstream basin and the outlet basin.

An integer vector of the new outflow indices adjusted for the sub-basin.

A list of integer vectors (`lst_Step_Cali`), where each element represents calibration cells at a specific step.

A list of integer vectors (`lst_Step_Cali`), where each element represents calibration cells at a specific step.
}
\description{
This function calculates inflow cells based on the outflow vector.

Creates a matrix of inflow cells for each cell.

Creates a list of inflow cells for each step.

Creates a field of last cells for each step.
}
\details{
routingtopology
}
